refactor(web): reduce settings-content to a lazy-loading shell
settings-content.tsx (4,891 lines) becomes a ~300-line shell that keeps only
routing, workspace selection, navigation, and the global save banner (via
useSettingsData). Every section is now loaded with next/dynamic, so the
production build emits one chunk per settings tab instead of a single bundle
that always contained all of them. TaskRunAnalyticsView keeps its previous
loading fallback text via the dynamic() loading option.

Also moves the (currently unreferenced) WebhooksSection into its own
sections/webhooks.tsx and escapes an apostrophe flagged by
react/no-unescaped-entities in github.tsx.
